Added more tests for relay convention limits.
